use parent folder for sibling games
