+++ src/components/games/remove-row-or-column/board-client.tsx
@@ -1,14 +1,13 @@
 import { useState } from 'react';
 import { range } from 'lodash';
 import {
-  strategyGameFactory, type BoardClientProps, type Events, type Ctx, GameBoard
+  type BoardClientProps, type Events, type Ctx, GameBoard
 } from '../../game-factory';
 import { useTranslation } from '../../language';
 import {
   type Board, type Orientation, type Move,
-  getRectangleAt, applyMove, isEmpty, generateStartBoard
+  getRectangleAt, applyMove, isEmpty
 } from './helpers';
-import { smartBotStrategy, randomBotStrategy } from './bot-strategy';
 
 type Selected = { r: number; c: number } | null
 
@@ -127,24 +126,6 @@ export const moves = {
   }
 };
 
-const rule = {
-  hu: <>
-    A játék kezdetén egy n × k-as téglalap minden mezőjére teszünk egy-egy korongot. A két játékos
-    felváltva lép. Egy lépésben a soron lévő játékos kiválaszt egy korongokból álló téglalapot, és
-    egy sorának vagy oszlopának minden korongját leveszi. (Korongokból álló téglalapnak egy olyan
-    téglalap alakú területet nevezünk, ahol minden mezőn van korong, de közvetlenül mellette sehol.
-    Kezdetben csak egy ilyen téglalap van, később már lehet, hogy több is.) Az nyer, aki az utolsó
-    korongot elveszi.
-  </>,
-  en: <>
-    At the start, a disc is placed on every cell of an n × k rectangle. The two players move
-    alternately. On a turn, the current player picks a solid rectangle of discs and removes every
-    disc in one of its rows or one of its columns. (A rectangle of discs is a rectangular block where
-    every cell has a disc and no cell directly next to it does. At the start there is only one such
-    rectangle; later there may be several.) Whoever removes the last disc wins.
-  </>
-};
-
 export const getPlayerStepDescription = ({ ctx }: { board: Board; ctx: Ctx }) => {
   return ctx.turnState
     ? {
@@ -156,22 +137,3 @@ export const getPlayerStepDescription = ({ ctx }: { board: Board; ctx: Ctx }) =>
       en: 'Click a disc, then choose to remove its row or its column.'
     };
 };
-
-export const RemoveRowOrColumn = strategyGameFactory({
-  presentation: {
-    rule,
-    getPlayerStepDescription
-  },
-  BoardClient,
-  gameplay: { moves },
-  variants: [
-    { botStrategy: randomBotStrategy, label: { hu: 'Teszt 🤖', en: 'Test 🤖' } },
-    // smart bot: optimal (Sprague–Grundy; moves to a zero position when winning)
-    {
-      botStrategy: smartBotStrategy,
-      generateStartBoard,
-      label: { hu: 'Okos 🤖', en: 'Smart 🤖' },
-      isDefault: true
-    }
-  ]
-});
