+++ src/components/games/remove-row-or-column/helpers.ts
@@ -1,5 +1,3 @@
-import { range, random, sample } from 'lodash';
-
 export type Grid = boolean[][]
 export type Board = { grid: Grid }
 export type Orientation = 'row' | 'col'
@@ -72,23 +70,4 @@ export const getAllMoves = (grid: Grid): Move[] => {
     for (let c = rect.minC; c <= rect.maxC; c++) moves.push({ r: rect.minR, c, orientation: 'col' });
   }
   return moves;
-};
-
-const makeFullGrid = (rows: number, cols: number): Grid =>
-  range(rows).map(() => range(cols).map(() => true));
-
-// Half the starting boards have both sides even (2nd player wins), half have an
-// odd side (1st player wins) — so each role wins ~50% across random starts.
-export const generateStartBoard = (): Board => {
-  let rows: number, cols: number;
-  if (random(0, 1)) {
-    rows = sample([2, 4, 6])!;
-    cols = sample([2, 4, 6])!;
-  } else {
-    do {
-      rows = sample([3, 4, 5, 6])!;
-      cols = sample([3, 4, 5, 6])!;
-    } while (rows % 2 === 0 && cols % 2 === 0);
-  }
-  return { grid: makeFullGrid(rows, cols) };
 };
